Only load explicitly enabled plugins
This introduces a new config value `PLUGINS_ENABLED`, which is a list of
plugin names. Those, and only those, which are configured in this list
will be loaded after they have been discovered in the plugins directory.

Use-case:

A container image can be produced, shipping multiple plugins, but still
giving users control over which ones get actually loaded.